LPD-46639 Move send email notification method from controller to marketplace service
