+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-etc-spring-boot/src/main/java/com/liferay/marketplace/TrialRestController.java
@@ -9,22 +9,17 @@ import com.liferay.client.extension.util.spring.boot3.BaseRestController;
 import com.liferay.client.extension.util.spring.boot3.LiferayOAuth2AccessTokenManager;
 import com.liferay.headless.admin.user.client.dto.v1_0.UserAccount;
 import com.liferay.headless.commerce.admin.order.client.dto.v1_0.Order;
-import com.liferay.headless.commerce.admin.order.client.pagination.Page;
 import com.liferay.headless.portal.instances.client.dto.v1_0.Admin;
 import com.liferay.headless.portal.instances.client.dto.v1_0.PortalInstance;
+import com.liferay.headless.portal.instances.client.pagination.Page;
 import com.liferay.headless.portal.instances.client.resource.v1_0.PortalInstanceResource;
 import com.liferay.marketplace.constants.MarketplaceConstants;
 import com.liferay.marketplace.service.ConsoleService;
 import com.liferay.marketplace.service.MarketplaceService;
-import com.liferay.notification.rest.client.dto.v1_0.NotificationQueueEntry;
-import com.liferay.notification.rest.client.dto.v1_0.NotificationTemplate;
-import com.liferay.notification.rest.client.resource.v1_0.NotificationQueueEntryResource;
-import com.liferay.notification.rest.client.resource.v1_0.NotificationTemplateResource;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portal.kernel.util.StringUtil;
 
 import java.net.URL;
 
@@ -72,8 +67,7 @@ public class TrialRestController extends BaseRestController {
 
 	@GetMapping("availability")
 	public String getAvailability() throws Exception {
-		com.liferay.headless.portal.instances.client.pagination.Page
-			<PortalInstance> page = _getPortalInstancesPage();
+		Page<PortalInstance> page = _getPortalInstancesPage();
 
 		return new JSONObject(
 		).put(
@@ -112,7 +106,7 @@ public class TrialRestController extends BaseRestController {
 		Map<String, String> customFields =
 			(Map<String, String>)order.getCustomFields();
 
-		_postNotificationQueueEntry(
+		_marketplaceService.postNotificationQueueEntry(
 			order.getCreatorEmailAddress(), "TRY-IT-NOW-EXPIRING-ORDER",
 			new HashMapBuilder<String, Object>().put(
 				"%TRIAL_CREATOR_FIRST_NAME%", userAccount.getGivenName()
@@ -150,8 +144,7 @@ public class TrialRestController extends BaseRestController {
 			_log.info("Provisioning order " + orderId);
 		}
 
-		com.liferay.headless.portal.instances.client.pagination.Page
-			<PortalInstance> portalInstancesPage = _getPortalInstancesPage();
+		Page<PortalInstance> portalInstancesPage = _getPortalInstancesPage();
 
 		if (portalInstancesPage.getTotalCount() == _TRIAL_MAX_INSTANCES) {
 			_log.error("Order is on hold");
@@ -190,7 +183,7 @@ public class TrialRestController extends BaseRestController {
 			"sendNotificationEmail", true);
 
 		if (sendNotificationEmail) {
-			_postNotificationQueueEntry(
+			_marketplaceService.postNotificationQueueEntry(
 				modelDTOOrderJSONObject.getString("creatorEmailAddress"),
 				"TRY-IT-NOW-PROCESSING-ORDER",
 				new HashMapBuilder<String, Object>().put(
@@ -233,7 +226,7 @@ public class TrialRestController extends BaseRestController {
 				orderId, MarketplaceConstants.ORDER_STATUS_IN_PROGRESS);
 
 			if (sendNotificationEmail) {
-				_postNotificationQueueEntry(
+				_marketplaceService.postNotificationQueueEntry(
 					modelDTOOrderJSONObject.getString("creatorEmailAddress"),
 					"TRY-IT-NOW-COMPLETED-ORDER",
 					new HashMapBuilder<String, Object>().put(
@@ -285,9 +278,8 @@ public class TrialRestController extends BaseRestController {
 		PortalInstanceResource portalInstanceResource =
 			_getPortalInstanceResource();
 
-		com.liferay.headless.portal.instances.client.pagination.Page
-			<PortalInstance> page =
-				portalInstanceResource.getPortalInstancesPage(true);
+		Page<PortalInstance> page =
+			portalInstanceResource.getPortalInstancesPage(true);
 
 		for (PortalInstance portalInstance : page.getItems()) {
 			if (Objects.equals(
@@ -318,112 +310,11 @@ public class TrialRestController extends BaseRestController {
 		).build();
 	}
 
-	private com.liferay.headless.portal.instances.client.pagination.Page
-		<PortalInstance> _getPortalInstancesPage() throws Exception {
-
+	private Page<PortalInstance> _getPortalInstancesPage() throws Exception {
 		PortalInstanceResource portalInstanceResource =
 			_getPortalInstanceResource();
 
 		return portalInstanceResource.getPortalInstancesPage(true);
-	}
-
-	private void _postNotificationQueueEntry(
-			String emailAddress, String externalReferenceCode,
-			Map<String, String> map)
-		throws Exception {
-
-		String authorization =
-			_liferayOAuth2AccessTokenManager.getAuthorization(
-				"liferay-marketplace-etc-spring-boot-oauth-application-" +
-					"headless-server");
-		URL liferayDXPURL = new URL(
-			lxcDXPServerProtocol + "://" + lxcDXPMainDomain);
-
-		NotificationTemplateResource notificationTemplateResource =
-			NotificationTemplateResource.builder(
-			).endpoint(
-				liferayDXPURL
-			).header(
-				HttpHeaders.AUTHORIZATION, authorization
-			).build();
-
-		NotificationTemplate notificationTemplate;
-
-		try {
-			notificationTemplate =
-				notificationTemplateResource.
-					getNotificationTemplateByExternalReferenceCode(
-						externalReferenceCode);
-		}
-		catch (Exception exception) {
-			_log.error(
-				"Unable to get notification template " + externalReferenceCode,
-				exception);
-
-			return;
-		}
-
-		NotificationQueueEntryResource notificationQueueEntryResource =
-			NotificationQueueEntryResource.builder(
-			).endpoint(
-				liferayDXPURL
-			).header(
-				HttpHeaders.AUTHORIZATION, authorization
-			).build();
-
-		NotificationQueueEntry notificationQueueEntry =
-			new NotificationQueueEntry();
-
-		notificationQueueEntry.setBody(
-			() -> _replace(
-				notificationTemplate.getBody(
-				).get(
-					"en_US"
-				),
-				map));
-
-		JSONArray jsonArray = new JSONObject(
-			String.valueOf(notificationTemplate)
-		).getJSONArray(
-			"recipients"
-		);
-
-		JSONObject jsonObject = jsonArray.getJSONObject(0);
-
-		notificationQueueEntry.setRecipients(
-			() -> new Object[] {
-				new HashMapBuilder<String, Object>().put(
-					"from", jsonObject.getString("from")
-				).put(
-					"fromName",
-					jsonObject.getJSONObject(
-						"fromName"
-					).getString(
-						"en_US"
-					)
-				).put(
-					"to", emailAddress
-				).build()
-			});
-
-		notificationQueueEntry.setSubject(
-			() -> _replace(
-				notificationTemplate.getSubject(
-				).get(
-					"en_US"
-				),
-				map));
-		notificationQueueEntry.setType(notificationTemplate::getType);
-
-		notificationQueueEntryResource.postNotificationQueueEntry(
-			notificationQueueEntry);
-
-		if (_log.isInfoEnabled()) {
-			_log.info(
-				StringBundler.concat(
-					"Sent ", externalReferenceCode, " notification to ",
-					emailAddress));
-		}
 	}
 
 	private PortalInstance _postPortalInstance(
@@ -464,15 +355,6 @@ public class TrialRestController extends BaseRestController {
 		}
 
 		return portalInstance;
-	}
-
-	private String _replace(String string, Map<String, String> map) {
-		for (Map.Entry<String, String> entry : map.entrySet()) {
-			string = StringUtil.replace(
-				string, entry.getKey(), entry.getValue());
-		}
-
-		return string;
 	}
 
 	private void _rollBackTrial(
